requirements-blocks.txt
matplotlib==3.1.0
numpy==1.16.4
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.0
